Refactor graph: All logic in UpdateGraphAroundActivity; use splines=ortho and newrank=true
