+++ src/de/hasait/cipa/activity/UpdateGraphAroundActivity.groovy
@@ -19,6 +19,7 @@ package de.hasait.cipa.activity
 import com.cloudbees.groovy.cps.NonCPS
 import de.hasait.cipa.Cipa
 import de.hasait.cipa.CipaInit
+import de.hasait.cipa.CipaRunContext
 import de.hasait.cipa.PScript
 import de.hasait.cipa.internal.CipaActivityWrapper
 
@@ -28,7 +29,10 @@ import java.util.regex.Pattern
 
 class UpdateGraphAroundActivity implements CipaInit, CipaAroundActivity, CipaAfterActivities, Serializable {
 
-	private final AtomicBoolean initCalled = new AtomicBoolean()
+	private final Map<CipaActivityWrapper, String> dotNodeNameByWrappers = new HashMap<>()
+
+	private final AtomicBoolean produceSVGCalled = new AtomicBoolean()
+
 	private final AtomicReference<String> svgContentHolder = new AtomicReference<>()
 
 	private Cipa cipa
@@ -53,11 +57,7 @@ class UpdateGraphAroundActivity implements CipaInit, CipaAroundActivity, CipaAft
 
 	@Override
 	void beforeActivityStarted(CipaActivityWrapper wrapper) {
-		if (!svgContentHolder.get()) {
-			if (initCalled.compareAndSet(false, true)) {
-				init()
-			}
-		}
+		// nop
 	}
 
 	@Override
@@ -82,55 +82,107 @@ class UpdateGraphAroundActivity implements CipaInit, CipaAroundActivity, CipaAft
 		updateGraph()
 	}
 
-	private void init() {
+	private String produceSVG() {
 		script.echo("Creating activity graph...")
+
+		String svgContent
 
 		script.dir(this.class.simpleName) {
 			String basename = 'activities-graph'
 			String dotFilename = "${basename}.graphviz.txt"
 			String svgFilename = "${basename}.svg"
-			script.writeFile(dotFilename, cipa.runContext.dotContent)
+			String dotContent = produceDot()
+			script.writeFile(dotFilename, dotContent)
 			script.sh("( test -x /usr/bin/dot && /usr/bin/dot -Tpng ${dotFilename} > ${basename}.png && /usr/bin/dot -Tsvg ${dotFilename} > ${svgFilename} ) || true")
 			script.archiveArtifacts("${basename}.*")
-			try {
-				String svgContent = script.readFile(svgFilename)
-				svgContentHolder.set(svgContent)
-			} catch (err) {
-				// ignore
-			}
+			svgContent = script.readFile(svgFilename)
 		}
 
-		updateGraph()
+		return svgContent
 	}
 
 	@NonCPS
-	private String produceSVG() {
-		String svgContent = svgContentHolder.get()
-		if (svgContent) {
-			for (wrapperWithNodeName in cipa.runContext.dotNodeNameByWrappers) {
-				CipaActivityWrapper wrapper = wrapperWithNodeName.key
-				String nodeName = wrapperWithNodeName.value
-				boolean running = wrapper.running
-				boolean failed = wrapper.failed
-				boolean depsFailed = wrapper.failedDependencies
-				boolean done = wrapper.done
-				boolean stable = wrapper.testResults.stable
-				String fill = depsFailed ? "gray" : (failed ? "red" : (done ? (stable ? "lightgreen" : "yellow") : "none"))
-				String stroke = running ? "blue" : "black"
-				String title = wrapper.activity.name.replace('<', '$lt;').replace('>', '$gt;')
-				svgContent = Pattern.compile("(<title>)${nodeName}(</title>\\s+<[a-z]+ fill=\")[^\"]+(\" stroke=\")[^\"]+(\")").matcher(svgContent).replaceFirst("\$1${title}\$2${fill}\$3${stroke}\$4")
+	private String produceDot() {
+		CipaRunContext runContext = cipa.runContext
+
+		StringBuilder dotContent = new StringBuilder()
+		dotContent << '\n'
+		dotContent << 'digraph pipeline {\n'
+		dotContent << 'rankdir="LR";\n'
+		dotContent << 'newrank=true;\n'
+		dotContent << 'splines=ortho;\n'
+		dotContent << 'node[shape=box];\n'
+		int activityI = 0
+
+		for (wrapper in runContext.wrappers) {
+			dotNodeNameByWrappers.put(wrapper, "a${activityI++}")
+		}
+		int nodeI = 0
+		for (node in runContext.nodes) {
+			dotContent << "subgraph cluster_node${nodeI++} {\n"
+			dotContent << "label=\"${node.label}\";\n"
+			dotContent << 'style=dotted;\n'
+			for (wrapper in runContext.wrappersByNode.get(node)) {
+				dotContent << "${dotNodeNameByWrappers.get(wrapper)}[label=\"${wrapper.activity.name}\"];\n"
+			}
+			dotContent << '}\n'
+		}
+		dotContent << 'start[shape=cds];\n'
+		for (wrapper in runContext.wrappers) {
+			Set<Map.Entry<CipaActivityWrapper, Boolean>> dependencies = wrapper.dependencies
+			if (dependencies.empty) {
+				dotContent << "start -> ${dotNodeNameByWrappers.get(wrapper)};\n"
+			} else {
+				for (dependency in dependencies) {
+					dotContent << "${dotNodeNameByWrappers.get(dependency.key)} -> ${dotNodeNameByWrappers.get(wrapper)}"
+					if (!dependency.value.booleanValue()) {
+						dotContent << ' [style=dashed]'
+					}
+					dotContent << ';\n'
+				}
 			}
 		}
+		dotContent << '}\n'
+		return dotContent.toString()
+	}
+
+	@NonCPS
+	private String transformSVG(String svgContent) {
+		for (wrapperWithNodeName in dotNodeNameByWrappers) {
+			CipaActivityWrapper wrapper = wrapperWithNodeName.key
+			String nodeName = wrapperWithNodeName.value
+			boolean running = wrapper.running
+			boolean failed = wrapper.failed
+			boolean depsFailed = wrapper.failedDependencies
+			boolean done = wrapper.done
+			boolean stable = wrapper.testResults.stable
+			String fill = depsFailed ? "gray" : (failed ? "red" : (done ? (stable ? "lightgreen" : "yellow") : "none"))
+			String stroke = running ? "blue" : "black"
+			String title = wrapper.activity.name.replace('<', '$lt;').replace('>', '$gt;')
+			svgContent = Pattern.compile("(<title>)${nodeName}(</title>\\s+<[a-z]+ fill=\")[^\"]+(\" stroke=\")[^\"]+(\")").matcher(svgContent).replaceFirst("\$1${title}\$2${fill}\$3${stroke}\$4")
+		}
+
 		return svgContent
 	}
 
 	private void updateGraph() {
-		String svgContent = produceSVG()
+		String svgContent = svgContentHolder.get()
+		if (!svgContent) {
+			if (produceSVGCalled.compareAndSet(false, true)) {
+				try {
+					svgContent = produceSVG()
+					svgContentHolder.set(svgContent)
+				} catch (err) {
+					script.echo("SVG creation failed: ${err}")
+				}
+			}
+		}
 		if (svgContent) {
+			svgContent = transformSVG(svgContent)
 			try {
 				script.rawScript.setCustomBuildProperty(key: "Activity-Graph", value: svgContent)
 			} catch (err) {
-				// ignore
+				script.echo("setCustomBuildProperty failed: ${err}")
 			}
 		}
 	}
